Faster unittesting
No need to test to see if saved animation has same number of frames as
expected. The more critical part about the animation function is that
the frames of the animation get saved. If the number of frames saved
to the output dir is as expected, then you can assume an animation
can be created from these. Reduced the number of frames from 5 to 2
since this is the minimum number of frames needed to create an
animation.