Jfrazier/reproduce smooth gww (#85)
# Summary

Now using ffmpeg to create the animation, more intelligent setting of the color limits for the pcolormesh emphasizes desired visual phenomena (e.g., gravity waves). Some example scripts that reproduce pretty animations are added in tests/exploratory/

# Details

* ffmpeg for animation, concat-dim time is what makes the difference here, using global max clim is required really... TODO: compare default xarray load with  time concat load

* TODO: need to use quantile/or provide user vmin/vmax

The clim is really important for showing phenomena. If you just use
the max over the entire time period, low velocity phenomena (which
domnates the phenomena) gets really washed out and is not visible.

The concat on the time dimension is correct. The other phenomena is not
correct.

* smooth figures for SAB... no feedback on this yet from C.stephan, but got feedback from y.morfa and p.ghosh, TODO: clean up quantile (i.e., make script arg) and clean up

* cleaning up CLI, added bluemarble to assets and updated default bluemarble paths, TODO: add quantile arg

* removed dead comments on using clim, removed use of masking

* added quantile argument

---------

Co-authored-by: Jared Frazier <frazier@iap-kborn.de>